Move isochrones query to `+page.svelte`
